[rilmodem] Do not remove reconnect timer
if reconnect timer exists, it means that ril_remove was called due
ril connection failure. In that case it should not be removed. If
reconnect timer does not exist, there is nothing to remove. Also
removing timer id for sim status retries in case of reconnect timer
is wrong either way. Calling of modem reset is unnecessary in boot.

Signed-off-by: Jussi Kangas <jussi.kangas@oss.tieto.com>